Create Fruit Shopping Calculator
This is a basic project made in 'Python' which take amount of money user have ,quantities of different fruits he wants to buy and then return the remaining amount of money left after shopping. Also, it tells that he can buy more items or not with left amount of money.